Only retry socket bind if returned EADDRINUSE
Currently, the NextRandomUdpSocket picks a random port in the range of
ephemeral ports and attempts to bind a UDP socket to it. When that bind
operation fails, it indiscriminately retries several times before
allowing the caller to poll again.

However, this means that if `bind` fails for another reason than the
requested port being unavailable, such as insufficient capabilities to
bind a socket, this error will not be propagated to the caller. Instead,
immediately return any error returned from `bind` unless it is
EADDRINUSE, in which case, pick another random port and try again as
before.